Standardize date display format across the application
Updates date formatting logic in `lib/utils.ts`, `BriefCard.tsx`, and `DownloadsScreen.tsx` to use "Month D, YYYY" format, and refactors `BriefDetailScreen.tsx` to display episode and summary metadata using consistent date and duration formats.

+++ client/components/BriefCard.tsx
@@ -32,7 +32,7 @@ function formatDuration(seconds: number): string {
 
 function formatDate(dateString: string): string {
   const d = new Date(dateString);
-  return d.toLocaleDateString("en-US", { month: "short", day: "numeric", year: "numeric" });
+  return d.toLocaleDateString("en-US", { month: "long", day: "numeric", year: "numeric" });
 }
 
 function getLanguageName(code: string): string {
